db will be filled for sure this time

+++ contemplor_server/app.js
@@ -1,3 +1,4 @@
+const server = require("./bin/www");
 var createError = require("http-errors");
 var express = require("express");
 var path = require("path");
@@ -53,13 +54,70 @@ app.use(function (err, req, res, next) {
   res.render("error");
 });
 
+async function waitForDatabase(maxRetries = 5, interval = 2000) {
+  let retries = 0;
+  while (retries < maxRetries) {
+    try {
+      await sequelize.authenticate();
+      console.log("Database connection has been established successfully.");
+      return true;
+    } catch (error) {
+      console.error(
+        `Unable to connect to the database (attempt ${retries + 1}):`,
+        error
+      );
+      retries++;
+      await new Promise((resolve) => setTimeout(resolve, interval));
+    }
+  }
+  return false;
+}
+
+function onError(error) {
+  if (error.syscall !== "listen") {
+    throw error;
+  }
+
+  var bind = typeof port === "string" ? "Pipe " + port : "Port " + port;
+
+  // handle specific listen errors with friendly messages
+  switch (error.code) {
+    case "EACCES":
+      console.error(bind + " requires elevated privileges");
+      process.exit(1);
+      break;
+    case "EADDRINUSE":
+      console.error(bind + " is already in use");
+      process.exit(1);
+      break;
+    default:
+      throw error;
+  }
+}
+
+function onListening() {
+  var addr = server.address();
+  var bind = typeof addr === "string" ? "pipe " + addr : "port " + addr.port;
+  debug("Listening on " + bind);
+}
+
 async function checkConnectionAndUpdateDatabase() {
   try {
-    await sequelize.authenticate();
-    console.log("Database connection has been established successfully.");
+    const connected = await waitForDatabase();
+    if (!connected) {
+      console.error(
+        "Failed to connect to the database after multiple attempts."
+      );
+      return;
+    }
 
     await sequelize.sync({ force: true }); // Dropping the table each time
     console.log("Database synchronized");
+
+    const port = server.address().port;
+    server.listen(port);
+    server.on("error", onError);
+    server.on("listening", onListening);
 
     const data1 = IndustryData.loadData();
     await IndustryData.saveData(data1);
